Simplify updating of GlobalTransform based on hierarchy of Parents
