+++ benchmarks/stress-test/src/lib.rs
@@ -20,7 +20,6 @@ use hotham::{
     systems::{
         animation_system, debug::debug_system, grabbing_system, hands_system, physics_system,
         rendering::rendering_system, skinning::skinning_system, update_global_transform_system,
-        update_global_transform_with_parent_system,
     },
     xr, Engine, HothamResult, TickData,
 };
@@ -250,7 +249,6 @@ fn tick(tick_props: &mut TickProps, tick_data: TickData) {
 
         animation_system(engine);
         update_global_transform_system(engine);
-        update_global_transform_with_parent_system(engine);
         skinning_system(engine);
     }
