+++ examples/crab-saber/src/lib.rs
@@ -8,7 +8,6 @@ use hotham::{
     systems::{
         audio_system, draw_gui_system, haptics_system, physics_system, pointers_system,
         rendering_system, update_global_transform_system,
-        update_global_transform_with_parent_system,
     },
     xr::{self, SessionState},
     Engine, HothamResult, TickData,
@@ -51,7 +50,6 @@ fn tick(tick_data: TickData, engine: &mut Engine, game_context: &mut GameContext
         game_system(engine, game_context);
 
         // Update world
-        update_global_transform_with_parent_system(engine);
         update_global_transform_system(engine);
 
         // Sync world with output contexts
