+++ examples/simple-scene/src/lib.rs
@@ -10,7 +10,7 @@ use hotham::{
     systems::{
         animation_system, debug::debug_system, grabbing_system, hands::add_hand, hands_system,
         physics_system, rendering::rendering_system, skinning::skinning_system,
-        update_global_transform_system, update_global_transform_with_parent_system,
+        update_global_transform_system,
     },
     xr, Engine, HothamResult, TickData,
 };
@@ -48,7 +48,6 @@ fn tick(tick_data: TickData, engine: &mut Engine, _state: &mut State) {
         physics_system(engine);
         animation_system(engine);
         update_global_transform_system(engine);
-        update_global_transform_with_parent_system(engine);
         skinning_system(engine);
         debug_system(engine);
     }
@@ -87,7 +86,6 @@ fn init(engine: &mut Engine) -> Result<(), hotham::HothamError> {
 
     // Update global transforms from local transforms before physics_system gets confused
     update_global_transform_system(engine);
-    update_global_transform_with_parent_system(engine);
 
     Ok(())
 }
