+++ hotham/src/systems/mod.rs
@@ -11,7 +11,6 @@ pub mod pointers;
 pub mod rendering;
 pub mod skinning;
 pub mod update_global_transform;
-pub mod update_global_transform_with_parent;
 
 pub use animation::animation_system;
 pub use audio::audio_system;
@@ -24,4 +23,3 @@ pub use pointers::pointers_system;
 pub use rendering::rendering_system;
 pub use skinning::skinning_system;
 pub use update_global_transform::update_global_transform_system;
-pub use update_global_transform_with_parent::update_global_transform_with_parent_system;
